Add README for Credential Manager, enhance platform exception handling, and improve response parsing
- Introduced a comprehensive README.md for the Credential Manager plugin, detailing setup, usage, and contribution guidelines.
- Enhanced platform exception handling by implementing a new utility class to convert platform exceptions into CredentialExceptions.
- Added a response parser utility to streamline the processing of credential responses from platform channels.
- Updated the credential_manager_core to register platform-specific plugins during initialization.
- Cleaned up and organized code across Android and iOS implementations for better maintainability.
