+++ packages/README.md
@@ -1,0 +1,50 @@
+# Credential Manager
+![Android](https://img.shields.io/badge/Platforms-Android-green)
+![iOS](https://img.shields.io/badge/Platforms-iOS-blue)
+
+[Credential Manager](https://developer.android.com/jetpack/androidx/releases/credentials) is a Jetpack API that supports multiple sign-in methods, such as username and password, passkeys, and federated sign-in solutions (like Sign-in with Google) in a single API, simplifying integration for developers on Android. For iOS, it uses Keychain for storing credentials.
+
+For users, Credential Manager unifies the sign-in interface across authentication methods, making it clearer and easier to sign into apps, regardless of the chosen method.
+
+For iOS, it uses [Keychain](https://developer.apple.com/documentation/security/keychain_services) and [Autofill](https://developer.apple.com/documentation/uikit/text_input/adding_password_autofill_support_to_your_app).
+
+For more information, please refer to the [official documentation](https://developer.android.com/jetpack/androidx/releases/credentials).
+
+## Getting Started
+
+Add the dependency to your pubspec.yaml file:
+
+```yaml
+credential_manager: <latest_version>
+```
+
+Or run:
+
+```bash
+flutter pub add credential_manager
+```
+
+## Documentation
+
+For comprehensive guides and detailed documentation, please visit our [Documentation Page](https://djsmk123.github.io/flutter_credential_manager_compose/#/). Here you will find everything you need to get started, including API references, usage examples, and best practices for integrating Credential Manager into your application.
+
+## Support the Project
+
+If you find this library useful and would like to support its development, consider buying me a coffee:
+
+[☕ Buy Me a Coffee](https://www.buymeacoffee.com/smkwinner)
+
+## Looking for Maintainers
+
+We are looking for maintainers to help us keep this library up-to-date and implement new features. If you are interested in contributing, please reach out via [GitHub](https://github.com/Djsmk123/flutter_credential_manager_compose) or open an issue. Your help would be greatly appreciated!
+
+## Contributors
+
+Thanks to all the contributors who have helped make this library better!
+
+[![Contributors](https://contrib.rocks/image?repo=Djsmk123/flutter_credential_manager_compose)](https://github.com/Djsmk123/flutter_credential_manager_compose/graphs/contributors)
+
+
+
+
+
