+++ packages/credential_manager/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"credential_manager_ios","path":"/Users/smkwinner/Desktop/workspace/flutter_credential_manager_compose/packages/credential_manager_ios/","native_build":true,"dependencies":[],"dev_dependency":false}],"android":[{"name":"credential_manager_android","path":"/Users/smkwinner/Desktop/workspace/flutter_credential_manager_compose/packages/credential_manager_android/","native_build":true,"dependencies":[],"dev_dependency":false}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"credential_manager_android","dependencies":[]},{"name":"credential_manager_ios","dependencies":[]}],"date_created":"2025-11-08 18:02:23.575272","version":"3.35.0","swift_package_manager_enabled":{"ios":false,"macos":false}}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"credential_manager_ios","path":"/Users/smkwinner/Desktop/workspace/flutter_credential_manager_compose/packages/credential_manager_ios/","native_build":true,"dependencies":[],"dev_dependency":false}],"android":[{"name":"credential_manager_android","path":"/Users/smkwinner/Desktop/workspace/flutter_credential_manager_compose/packages/credential_manager_android/","native_build":true,"dependencies":[],"dev_dependency":false}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"credential_manager_android","dependencies":[]},{"name":"credential_manager_ios","dependencies":[]}],"date_created":"2025-11-08 18:20:52.969516","version":"3.35.0","swift_package_manager_enabled":{"ios":false,"macos":false}}
