+++ packages/credential_manager/example/ios/Runner/Runner.entitlements
@@ -2,11 +2,6 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-	<key>com.apple.developer.associated-domains</key>
-	<array>
-		<string>webcredentials:blogs-deeplink-example.vercel.app?mode=developer</string>
-		<string>applinks:blogs-deeplink-example.vercel.app</string>
-	</array>
 	<key>keychain-access-groups</key>
 	<array>
 		<string>$(AppIdentifierPrefix)com.smkwinner.credential-manager-example</string>
