+++ packages/credential_manager/lib/src/utils/credential_type.dart
@@ -1,0 +1,7 @@
+/// Enum representing the different types of credentials.
+enum CredentialType {
+  passwordCredentials,
+  publicKeyCredentials,
+  googleIdTokenCredentials
+}
+
