+++ packages/credential_manager_android/lib/credential_manager_android.dart
@@ -1,4 +1,5 @@
 library credential_manager_android;
 
 export 'src/credential_manager_android.dart';
+export 'src/credential_manager_android_plugin.dart';
