+++ packages/credential_manager_ios/lib/credential_manager_ios.dart
@@ -1,4 +1,5 @@
 library credential_manager_ios;
 
 export 'src/credential_manager_ios.dart';
+export 'src/credential_manager_ios_plugin.dart';
